+++ ChangeLog.md
@@ -3,7 +3,8 @@
 ## 2.1.0 (unreleased)
 
 ### Features
-- Add a custom tool analyzer (`CTRL+T`) to compare two files, e.g. run `otool -L` on both files and show the `diff -u` between them.
+- Add a custom tool analyzer (`SHIFT+T`) to compare two files, e.g. run `otool -L` on both files and show the `diff -u` between them.
+- Add selected file (`SHIFT+J`) or all files (`SHIFT+I`) identification
 
 ## 2.0.1 (8 Feb 2022)